requirements.txt
flask==3.0.0
flask-cors==4.0.0
PyYAML==6.0.1
gunicorn==21.2.0
python-dotenv==1.0.0